REFACTOR: (SPOT-17) 중간지점 계산, 경로 조회 서비스를 implement 레이어로 옮긴다
