+++ src/main/java/com/meetup/server/event/implement/MeetingPointCalculator.java
@@ -1,11 +1,9 @@
-package com.meetup.server.event.application;
+package com.meetup.server.event.implement;
 
 import com.meetup.server.event.domain.Event;
 import com.meetup.server.event.dto.response.MeetingPointResult;
 import com.meetup.server.event.exception.EventErrorType;
 import com.meetup.server.event.exception.EventException;
-import com.meetup.server.event.implement.EventReader;
-import com.meetup.server.event.implement.EventValidator;
 import com.meetup.server.global.util.CoordinateUtil;
 import com.meetup.server.startpoint.domain.StartPoint;
 import com.meetup.server.startpoint.implement.StartPointReader;
@@ -16,17 +14,17 @@ import com.meetup.server.subway.implement.reader.SubwayReader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.locationtech.jts.geom.Point;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
-@Service
-@RequiredArgsConstructor
 @Slf4j
-public class MeetingPointService {
+@Component
+@RequiredArgsConstructor
+public class MeetingPointCalculator {
 
     private final EventReader eventReader;
     private final StartPointReader startPointReader;
@@ -35,7 +33,7 @@ public class MeetingPointService {
     private final EventValidator eventValidator;
 
     @Transactional
-    public List<MeetingPointResult> getMeetingPoints(UUID eventId) {
+    public List<MeetingPointResult> calculate(UUID eventId) {
         Event event = eventReader.read(eventId);
         List<StartPoint> startPoints = startPointReader.readAll(event);
         eventValidator.validateMinimumStartPoints(startPoints);
