+++ src/main/java/com/meetup/server/event/implement/RouteAssembler.java
@@ -1,4 +1,4 @@
-package com.meetup.server.event.application;
+package com.meetup.server.event.implement;
 
 import com.meetup.server.event.domain.Event;
 import com.meetup.server.event.dto.response.MeetingPointRouteGroup;
@@ -10,24 +10,23 @@ import com.meetup.server.startpoint.implement.StartPointReader;
 import com.meetup.server.subway.domain.Subway;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Slf4j
-@Service
+@Component
 @RequiredArgsConstructor
-public class RouteService {
+public class RouteAssembler {
 
     private final ParkingLotFinder parkingLotFinder;
-    private final RouteDetailService routeDetailService;
+    private final RouteDetailFetcher routeDetailFetcher;
     private final StartPointReader startPointReader;
 
-    public MeetingPointRouteGroup getAllRouteDetails(Event event, List<StartPoint> startPointList, Subway subway) {
-
+    public MeetingPointRouteGroup assemble(Event event, List<StartPoint> startPointList, Subway subway) {
         List<RouteResponse> routeList = startPointList.stream()
-                .map(startPoint -> routeDetailService.fetchPerRouteDetails(
+                .map(startPoint -> routeDetailFetcher.fetch(
                         startPoint,
                         String.valueOf(startPoint.getLocation().getRoadLongitude()),
                         String.valueOf(startPoint.getLocation().getRoadLatitude()),
@@ -37,7 +36,6 @@ public class RouteService {
                 .collect(Collectors.toList());
 
         ClosestParkingLot closestParkingLot = parkingLotFinder.findClosestParkingLot(subway.getPoint());
-
         StartPoint earliestStartPoint = startPointReader.readEarliestByEventId(event.getEventId());
         return MeetingPointRouteGroup.of(earliestStartPoint, routeList, subway, closestParkingLot);
     }
